backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
PyJWT
python-multipart
email-validator
reportlab
python-dotenv
